+++ modules/darwin-settings/default.nix
@@ -3,7 +3,6 @@
   imports = [
     # ./custom-system-preferences.nix
     ./dock.nix
-    ./documentation.nix
     ./finder.nix
     ./homebrew.nix
     ./keyboard.nix