+++ overlays/default.nix
@@ -167,6 +167,11 @@
     nixt = prev.nixt.override { nix = prev.nixVersions.latest; };
   };
 
+  nix-monitored = _: prev: {
+    nix-monitored =
+      self.inputs.nix-monitored.packages.${prev.stdenv.system}.default;
+  };
+
   # See also: https://github.com/BKSalman/nix_config/commit/8d94944af411bfff74edafce18ea1d0ca4789bb9
   ranger = _: prev: {
     ranger = prev.ranger.overrideAttrs (old: {